trocando o tempo para float

+++ src/main/java/simpleapi/dao/TimerDao.java
@@ -4,6 +4,6 @@ import simpleapi.domain.SkillTimer;
 import simpleapi.exception.NotFoundException;
 
 public interface TimerDao {
-	public long getTime(int id) throws NotFoundException;
+	public float getTime(int id) throws NotFoundException;
 	public void saveTime(SkillTimer timer);
 }
